refactor: 마지막 돌의 위치를 반환 기능 책임 위임
- GoStone에서 Board로 위임
