+++ src/main/kotlin/omok/model/GoStone.kt
@@ -2,22 +2,20 @@ package omok.model
 
 abstract class GoStone {
     abstract val stoneType: Stone
-    private lateinit var lastPosition: Position
 
-    fun getLastPosition(): Position? {
-        if (::lastPosition.isInitialized) return lastPosition
-        return null
+    fun putStone(position: Position): Stone {
+        require(Board.board[position.col][position.row] == Stone.NONE) { "이미 놓여진 자리입니다." }
+        Board.board[position.col][position.row] = stoneType
+        Board.lastPosition = position
+        if (stoneType == Stone.BLACK_STONE) {
+            return Stone.WHITE_STONE
+        }
+        return Stone.BLACK_STONE
     }
 
-    fun putStone(position: Position) {
-        require(Board.board[position.row][position.col] == Stone.NONE) { "이미 놓여진 자리입니다." }
-        Board.board[position.row][position.col] = stoneType
-        lastPosition = position
-    }
+    fun findOmok(position: Position): Boolean {
+        if (Board.board[position.col][position.row] != stoneType) return false
 
-    fun findOmok(): Boolean {
-        if (Board.board[lastPosition.row][lastPosition.col] != stoneType) return false
-
-        return Board.findOmok(lastPosition, stoneType)
+        return Board.findOmok(position, stoneType)
     }
 }
