+++ src/test/kotlin/omok/model/BlackStoneTest.kt
@@ -7,23 +7,6 @@ import org.junit.jupiter.api.assertDoesNotThrow
 
 class BlackStoneTest {
     @Test
-    fun `마지막 돌이 없다면 null을 반환한다`() {
-        val stone = BlackStone()
-        val actual = stone.getLastPosition()
-        val expected = null
-        assertThat(actual).isEqualTo(expected)
-    }
-
-    @Test
-    fun `마지막 돌이 있다면 돌의 위치를 반환한다`() {
-        val stone = BlackStone()
-        stone.putStone(Position(1, 5))
-        val actual = stone.getLastPosition()
-        val expected = Position(1, 5)
-        assertThat(actual).isEqualTo(expected)
-    }
-
-    @Test
     fun `이미 돌이 놓인 자리에 돌을 놓을 경우 예외가 발생한다`() {
         val stone = BlackStone()
         val position = Position(0, 0)
@@ -51,8 +34,9 @@ class BlackStoneTest {
         blackStone.putStone(Position(4, 4))
         blackStone.putStone(Position(5, 5))
 
+        val lastPosition = Position(4, 4)
         // when
-        val actual = blackStone.findOmok()
+        val actual = blackStone.findOmok(lastPosition)
         val expected = true
 
         // then
@@ -70,8 +54,9 @@ class BlackStoneTest {
         blackStone.putStone(Position(13, 13))
         blackStone.putStone(Position(14, 14))
 
+        val lastPosition = Position(11, 11)
         // when
-        val actual = blackStone.findOmok()
+        val actual = blackStone.findOmok(lastPosition)
         val expected = false
 
         // then
